some frustration

+++ parse_opencv_output.py
@@ -19,7 +19,7 @@ def main():
             query_y = float(tokens[10])
             train_x = float(tokens[17])
             train_y = float(tokens[18])
-            distance = int(tokens[20])
+            distance = float(tokens[20])
             output = output + "{0} {1} {2} {3} {4}\n".format(query_x, query_y, train_x, train_y, distance)
     print(output)
     with open("parsed_output.txt", "w") as text_file:
